Move connection out of execution
