+++ src/test/datascience/execution.unit.test.ts
@@ -60,10 +60,8 @@ import { INotebookServerFactory } from '../../kernels/jupyter/types';
 import { IJupyterSubCommandExecutionService } from '../../kernels/jupyter/types.node';
 import { SystemVariables } from '../../platform/common/variables/systemVariables.node';
 import { getOSType, OSType } from '../../platform/common/utils/platform';
-import { JupyterUriProviderRegistration } from '../../kernels/jupyter/jupyterUriProviderRegistration';
-import { JupyterSessionManagerFactory } from '../../kernels/jupyter/session/jupyterSessionManagerFactory';
 import { JupyterServerUriStorage } from '../../kernels/jupyter/launcher/serverUriStorage';
-import { ServerConnectionType } from '../../kernels/jupyter/launcher/serverConnectionType';
+import { JupyterConnection } from '../../kernels/jupyter/jupyterConnection';
 
 /* eslint-disable @typescript-eslint/no-explicit-any, , no-multi-str,  */
 class DisposableRegistry implements IAsyncDisposableRegistry {
@@ -86,7 +84,7 @@ class DisposableRegistry implements IAsyncDisposableRegistry {
     };
 }
 
-suite('Jupyter Execution', async () => {
+suite.only('Jupyter Execution', async () => {
     const interpreterService = mock<IInterpreterService>();
     const jupyterOutputChannel = new MockOutputChannel('');
     const executionFactory = mock(PythonExecutionFactory);
@@ -1000,15 +998,9 @@ suite('Jupyter Execution', async () => {
         when(kernelFinder.listKernels(anything(), anything())).thenResolve([kernelMetadata]);
         when(serviceContainer.get<NotebookStarter>(NotebookStarter)).thenReturn(notebookStarter);
         when(serviceContainer.get<ILocalKernelFinder>(ILocalKernelFinder)).thenReturn(instance(kernelFinder));
-        const sessionManagerFactory = mock(JupyterSessionManagerFactory);
         const serverFactory = mock<INotebookServerFactory>();
-        const provider = mock(JupyterUriProviderRegistration);
         const serverUriStorage = mock(JupyterServerUriStorage);
-        const connectionType = mock<ServerConnectionType>();
-        when(connectionType.isLocalLaunch).thenReturn(true);
-        const onDidChangeEvent = new EventEmitter<void>();
-        disposableRegistry.push(onDidChangeEvent);
-        when(connectionType.onDidChange).thenReturn(onDidChangeEvent.event);
+        const connection = mock<JupyterConnection>();
         return {
             executionService: activeService.object,
             jupyterExecution: new HostJupyterExecution(
@@ -1019,11 +1011,9 @@ suite('Jupyter Execution', async () => {
                 instance(configService),
                 notebookStarter,
                 jupyterCmdExecutionService,
-                instance(provider),
-                instance(sessionManagerFactory),
                 instance(serverFactory),
                 instance(serverUriStorage),
-                instance(connectionType)
+                instance(connection)
             )
         };
     }
